+++ plugins/supabase/iac/index.spec.ts
@@ -1354,7 +1354,7 @@ describe("resolveEdgeFunctionDenoConfig", () => {
         mime: `npm:mime@${resolvePackageVersion("mime", {
           searchFrom: path.resolve("plugins/plugin-core"),
         })}`,
-        semver: `npm:semver@${resolvePackageVersion("semver", {
+        verkit: `npm:verkit@${resolvePackageVersion("verkit", {
           searchFrom: path.resolve("plugins/plugin-core"),
         })}`,
       });